backend/requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
qdrant-client>=1.7.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
